+++ src/core/context.cpp
@@ -1,9 +1,9 @@
 #include "pch.h"
 #include "context.h"
-#include "utility.h"
 #include "conf.h"
-#include "tasks/task.h"
-#include "tools/tools.h"
+#include "../utility.h"
+#include "../tasks/task.h"
+#include "../tools/tools.h"
 
 namespace mob::details
 {
