+++ src/core/op.cpp
@@ -1,10 +1,10 @@
 #include "pch.h"
 #include "op.h"
-#include "utility.h"
 #include "conf.h"
 #include "context.h"
 #include "process.h"
-#include "tools/tools.h"
+#include "../utility.h"
+#include "../tools/tools.h"
 
 namespace mob::op
 {
