+++ src/core/pch.h
@@ -1,0 +1,2 @@
+// for intellisense
+#include "../pch.h"
