+++ src/core/process.cpp
@@ -1,9 +1,9 @@
 #include "pch.h"
 #include "process.h"
 #include "conf.h"
-#include "net.h"
 #include "context.h"
 #include "op.h"
+#include "../net.h"
 
 namespace mob
 {
