+++ src/core/process.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "utility.h"
 #include "env.h"
 #include "context.h"
+#include "../utility.h"
 
 namespace mob
 {
