+++ src/utility.cpp
@@ -1,10 +1,10 @@
 #include "pch.h"
 #include "utility.h"
-#include "conf.h"
-#include "op.h"
 #include "net.h"
-#include "process.h"
-#include "context.h"
+#include "core/conf.h"
+#include "core/op.h"
+#include "core/process.h"
+#include "core/context.h"
 
 namespace mob
 {
