+++ vs/mob.vcxproj
@@ -59,17 +59,17 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\src\commands.cpp" />
-    <ClCompile Include="..\src\conf.cpp" />
-    <ClCompile Include="..\src\context.cpp" />
-    <ClCompile Include="..\src\env.cpp" />
+    <ClCompile Include="..\src\core\conf.cpp" />
+    <ClCompile Include="..\src\core\context.cpp" />
+    <ClCompile Include="..\src\core\env.cpp" />
+    <ClCompile Include="..\src\core\op.cpp" />
+    <ClCompile Include="..\src\core\process.cpp" />
     <ClCompile Include="..\src\main.cpp" />
     <ClCompile Include="..\src\net.cpp" />
-    <ClCompile Include="..\src\op.cpp" />
     <ClCompile Include="..\src\pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="..\src\process.cpp" />
     <ClCompile Include="..\src\tasks\boost.cpp" />
     <ClCompile Include="..\src\tasks\boost_di.cpp" />
     <ClCompile Include="..\src\tasks\bzip2.cpp" />
@@ -109,13 +109,13 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\commands.h" />
-    <ClInclude Include="..\src\conf.h" />
-    <ClInclude Include="..\src\context.h" />
-    <ClInclude Include="..\src\env.h" />
-    <ClInclude Include="..\src\op.h" />
+    <ClInclude Include="..\src\core\conf.h" />
+    <ClInclude Include="..\src\core\context.h" />
+    <ClInclude Include="..\src\core\env.h" />
+    <ClInclude Include="..\src\core\op.h" />
+    <ClInclude Include="..\src\core\process.h" />
     <ClInclude Include="..\src\net.h" />
     <ClInclude Include="..\src\pch.h" />
-    <ClInclude Include="..\src\process.h" />
     <ClInclude Include="..\src\tasks\task.h" />
     <ClInclude Include="..\src\tasks\tasks.h" />
     <ClInclude Include="..\src\tools\tools.h" />
